Moved kdf out of the node
So the contract code can inherit it (also maybe include this in a client
library later?)
