+++ crypto-shared/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "crypto-shared"
+edition = "2021"
+
+[dependencies]
+k256 = { version = "0.13.1", features = ["sha256", "ecdsa", "serde", "arithmetic", "expose-field"] }
+hkdf = "0.12.4"
+anyhow = "1"
+near-primitives = "0.17.0"
