+++ crypto-shared/src/lib.rs
@@ -1,0 +1,5 @@
+mod kdf;
+mod types;
+
+pub use kdf::{derive_epsilon, derive_key, into_eth_sig, x_coordinate};
+pub use types::{PublicKey, ScalarExt};
