+++ node/Cargo.toml
@@ -51,6 +51,7 @@ near-sdk = { version = "=5.1.0", features = ["legacy", "unit-testing"] }
 
 mpc-contract = { path = "../contract" }
 mpc-keys = { path = "../keys" }
+crypto-shared = { path = "../crypto-shared" }
 
 itertools = "0.12.0"
 prometheus = { version = "0.13.3" }
