+++ node/src/protocol/state.rs
@@ -6,8 +6,9 @@ use super::triple::TripleManager;
 use super::SignQueue;
 use crate::http_client::MessageQueue;
 use crate::storage::triple_storage::TripleData;
-use crate::types::{KeygenProtocol, PublicKey, ReshareProtocol, SecretKeyShare};
+use crate::types::{KeygenProtocol, ReshareProtocol, SecretKeyShare};
 use cait_sith::protocol::Participant;
+use crypto_shared::PublicKey;
 use near_account_id::AccountId;
 use serde::{Deserialize, Serialize};
 use std::fmt;
